chore: rename TranslateMenu to NavigationAppBarMenuLanguage
